Refactoring code, and adding more tests based on feedback
