+++ frontend/app-development/layout/SettingsModalButton/SettingsModal/SettingsModal.tsx
@@ -1,6 +1,6 @@
 import React, { ReactNode, useState } from 'react';
 import classes from './SettingsModal.module.css';
-import { Alert, ErrorMessage, Heading, Paragraph, Spinner } from '@digdir/design-system-react';
+import { Heading } from '@digdir/design-system-react';
 import {
   CogIcon,
   InformationSquareIcon,
@@ -15,17 +15,9 @@ import { SettingsModalTab } from 'app-development/types/SettingsModalTab';
 import { createNavigationTab } from './utils';
 import { useTranslation } from 'react-i18next';
 import { PolicyTab } from './components/Tabs/PolicyTab';
-import { AboutTab } from './components/Tabs/AbouTab';
+import { AboutTab } from './components/Tabs/AboutTab';
 import { LocalChangesTab } from './components/Tabs/LocalChangesTab';
 import { AccessControlTab } from './components/Tabs/AccessControlTab';
-import {
-  useAppPolicyQuery,
-  useAppConfigQuery,
-  useAppMetadataQuery,
-} from 'app-development/hooks/queries';
-import { useRepoInitialCommitQuery, useRepoMetadataQuery } from 'app-shared/hooks/queries';
-import { Center } from 'app-shared/components/Center';
-import { mergeQueryStatuses } from 'app-shared/utils/tanstackQueryUtils';
 
 export type SettingsModalProps = {
   /**
@@ -60,32 +52,6 @@ export type SettingsModalProps = {
  */
 export const SettingsModal = ({ isOpen, onClose, org, app }: SettingsModalProps): ReactNode => {
   const { t } = useTranslation();
-
-  const {
-    status: policyStatus,
-    data: policyData,
-    error: policyError,
-  } = useAppPolicyQuery(org, app);
-  const {
-    status: appConfigStatus,
-    data: appConfigData,
-    error: appConfigError,
-  } = useAppConfigQuery(org, app);
-  const {
-    status: repositoryStatus,
-    data: repositoryData,
-    error: repositoryError,
-  } = useRepoMetadataQuery(org, app);
-  const {
-    status: initialCommitStatus,
-    data: initialCommitData,
-    error: initialCommitError,
-  } = useRepoInitialCommitQuery(org, app);
-  const {
-    status: appMetadataStatus,
-    data: appMetadataData,
-    error: appMetadataError,
-  } = useAppMetadataQuery(org, app);
 
   const [currentTab, setCurrentTab] = useState<SettingsModalTab>('about');
 
@@ -142,78 +108,16 @@ export const SettingsModal = ({ isOpen, onClose, org, app }: SettingsModalProps)
   const displayTabs = () => {
     switch (currentTab) {
       case 'about': {
-        return (
-          <AboutTab
-            appConfig={appConfigData}
-            org={org}
-            app={app}
-            repository={repositoryData}
-            createdBy={initialCommitData.author.name}
-          />
-        );
+        return <AboutTab org={org} app={app} />;
       }
       case 'policy': {
-        return <PolicyTab policy={policyData} org={org} app={app} />;
+        return <PolicyTab org={org} app={app} />;
       }
       case 'accessControl': {
-        return <AccessControlTab appMetadata={appMetadataData} org={org} app={app} />;
+        return <AccessControlTab org={org} app={app} />;
       }
       case 'localChanges': {
         return <LocalChangesTab org={org} app={app} />;
-      }
-    }
-  };
-
-  /**
-   * Based on the state of the API calls, display spinner, error or components
-   */
-  const displayModalContent = () => {
-    switch (
-      mergeQueryStatuses(
-        policyStatus,
-        appConfigStatus,
-        appMetadataStatus,
-        repositoryStatus,
-        initialCommitStatus,
-      )
-    ) {
-      case 'loading': {
-        return (
-          <div className={classes.modalContent}>
-            <Center>
-              <Spinner
-                size='2xLarge'
-                variant='interaction'
-                title={t('settings_modal.loading_content')}
-              />
-            </Center>
-          </div>
-        );
-      }
-      case 'error': {
-        return (
-          <Center>
-            <Alert severity='danger'>
-              <Paragraph>{t('general.fetch_error_message')}</Paragraph>
-              <Paragraph>{t('general.error_message_with_colon')}</Paragraph>
-              {policyError && <ErrorMessage>{policyError.message}</ErrorMessage>}
-              {appConfigError && <ErrorMessage>{appConfigError.message}</ErrorMessage>}
-              {repositoryError && <ErrorMessage>{repositoryError.message}</ErrorMessage>}
-              {initialCommitError && <ErrorMessage>{initialCommitError.message}</ErrorMessage>}
-              {appMetadataError && <ErrorMessage>{appMetadataError.message}</ErrorMessage>}
-            </Alert>
-          </Center>
-        );
-      }
-      case 'success': {
-        return (
-          <div className={classes.modalContent}>
-            <div className={classes.leftNavWrapper}>
-              <LeftNavigationBar tabs={leftNavigationTabs} className={classes.leftNavigationBar} />
-            </div>
-            <div className={classes.tabWrapper}>{displayTabs()}</div>
-          </div>
-        );
       }
     }
   };
@@ -231,7 +135,12 @@ export const SettingsModal = ({ isOpen, onClose, org, app }: SettingsModalProps)
         </div>
       }
     >
-      {displayModalContent()}
+      <div className={classes.modalContent}>
+        <div className={classes.leftNavWrapper}>
+          <LeftNavigationBar tabs={leftNavigationTabs} className={classes.leftNavigationBar} />
+        </div>
+        <div className={classes.tabWrapper}>{displayTabs()}</div>
+      </div>
     </Modal>
   );
 };
