+++ frontend/app-development/layout/SettingsModalButton/SettingsModal/components/LoadingTabData/index.ts
@@ -1,0 +1,1 @@
+export { LoadingTabData } from './LoadingTabData';
