+++ frontend/app-development/layout/SettingsModalButton/SettingsModal/components/TabDataError/index.ts
@@ -1,0 +1,1 @@
+export { TabDataError } from './TabDataError';
