+++ frontend/app-development/layout/SettingsModalButton/SettingsModal/components/Tabs/AboutTab/CreatedFor/CreatedFor.module.css
@@ -1,4 +1,4 @@
-.contentWrapper {
+.wrapper {
   max-width: 500px;
 }
