+++ frontend/app-development/layout/SettingsModalButton/SettingsModal/components/Tabs/AboutTab/CreatedFor/index.ts
@@ -1,0 +1,1 @@
+export { CreatedFor } from './CreatedFor';
