+++ frontend/app-development/layout/SettingsModalButton/SettingsModal/components/Tabs/AboutTab/InputFields/InputFields.module.css
@@ -1,0 +1,3 @@
+.wrapper {
+  max-width: 500px;
+}
