+++ frontend/app-development/layout/SettingsModalButton/SettingsModal/components/Tabs/AboutTab/InputFields/index.ts
@@ -1,0 +1,1 @@
+export { InputFields } from './InputFields';
