+++ frontend/app-development/layout/SettingsModalButton/SettingsModal/components/Tabs/AboutTab/index.ts
