+++ frontend/app-development/layout/SettingsModalButton/SettingsModal/components/Tabs/AccessControlTab/AccessControlTab.test.tsx
@@ -1,5 +1,10 @@
 import React from 'react';
-import { act, render as rtlRender, screen } from '@testing-library/react';
+import {
+  act,
+  render as rtlRender,
+  screen,
+  waitForElementToBeRemoved,
+} from '@testing-library/react';
 import { AccessControlTab, AccessControlTabProps } from './AccessControlTab';
 import { textMock } from '../../../../../../../testing/mocks/i18nMock';
 import { createQueryClientMock } from 'app-shared/mocks/queryClientMock';
@@ -23,30 +28,56 @@ mockUpdateAppMetadataMutation.mockReturnValue({
   mutate: updateAppMetadataMutation,
 } as unknown as UseMutationResult<void, unknown, ApplicationMetadata, unknown>);
 
-describe('AccessControlTab', () => {
-  const user = userEvent.setup();
+const getAppMetadata = jest.fn().mockImplementation(() => Promise.resolve({}));
 
+const defaultProps: AccessControlTabProps = {
+  org: mockOrg,
+  app: mockApp,
+};
+
+describe('AccessControlTab', () => {
   afterEach(jest.clearAllMocks);
 
-  const defaultProps: AccessControlTabProps = {
-    appMetadata: mockAppMetadata,
-    org: mockOrg,
-    app: mockApp,
-  };
+  it('initially displays the spinner when loading data', () => {
+    render();
 
-  it('should render all checkboxes as unchecked when applicationMetadata des not contain partyTypes allowed', () => {
-    render({}, createQueryClientMock(), {
-      ...defaultProps,
-      appMetadata: { ...mockAppMetadata, partyTypesAllowed: null },
-    });
+    expect(screen.getByTitle(textMock('settings_modal.loading_content'))).toBeInTheDocument();
+  });
+
+  it('fetches appMetadata on mount', () => {
+    render();
+    expect(getAppMetadata).toHaveBeenCalledTimes(1);
+  });
+
+  it('shows an error message if an error occured on the getAppMetadata query', async () => {
+    const errorMessage = 'error-message-test';
+    render({}, { getAppMetadata: () => Promise.reject({ message: errorMessage }) });
+
+    await waitForElementToBeRemoved(() =>
+      screen.queryByTitle(textMock('settings_modal.loading_content')),
+    );
+
+    expect(screen.getByText(textMock('general.fetch_error_message'))).toBeInTheDocument();
+    expect(screen.getByText(textMock('general.error_message_with_colon'))).toBeInTheDocument();
+    expect(screen.getByText(errorMessage)).toBeInTheDocument();
+  });
+
+  it('should render all checkboxes as unchecked when applicationMetadata des not contain partyTypes allowed', async () => {
+    getAppMetadata.mockImplementation(() =>
+      Promise.resolve({ ...mockAppMetadata, partyTypesAllowed: null }),
+    );
+    render();
+    await waitForElementToBeRemoved(() =>
+      screen.queryByTitle(textMock('settings_modal.loading_content')),
+    );
 
     const checkboxes = screen.queryAllByRole('checkbox');
     expect(checkboxes).toHaveLength(4);
     checkboxes.forEach((c) => expect(c).not.toBeChecked());
   });
 
-  it('should render all checkboxes with the correct values based on the party types allowed', () => {
-    render({}, createQueryClientMock(), defaultProps);
+  it('should render all checkboxes with the correct values based on the party types allowed', async () => {
+    await resolveAndWaitForSpinnerToRemove();
 
     const bankruptcyEstateCheckbox = screen.getByRole('checkbox', {
       name: textMock('settings_modal.access_control_tab_option_bankruptcy_estate'),
@@ -70,7 +101,8 @@ describe('AccessControlTab', () => {
   });
 
   it('handles checkbox changes', async () => {
-    render({}, createQueryClientMock(), defaultProps);
+    const user = userEvent.setup();
+    await resolveAndWaitForSpinnerToRemove();
 
     const organisationCheckboxBefore = screen.getByRole('checkbox', {
       name: textMock('settings_modal.access_control_tab_option_organisation'),
@@ -86,7 +118,8 @@ describe('AccessControlTab', () => {
   });
 
   it('calles saving function when checboxgroup is blurred', async () => {
-    render({}, createQueryClientMock(), defaultProps);
+    const user = userEvent.setup();
+    await resolveAndWaitForSpinnerToRemove();
 
     const organisationCheckboxBefore = screen.getByRole('checkbox', {
       name: textMock('settings_modal.access_control_tab_option_organisation'),
@@ -98,19 +131,28 @@ describe('AccessControlTab', () => {
   });
 });
 
+const resolveAndWaitForSpinnerToRemove = async (props: Partial<AccessControlTabProps> = {}) => {
+  getAppMetadata.mockImplementation(() => Promise.resolve(mockAppMetadata));
+  render(props);
+  await waitForElementToBeRemoved(() =>
+    screen.queryByTitle(textMock('settings_modal.loading_content')),
+  );
+};
+
 const render = (
+  props: Partial<AccessControlTabProps> = {},
   queries: Partial<ServicesContextProps> = {},
   queryClient: QueryClient = createQueryClientMock(),
-  props: AccessControlTabProps,
 ) => {
   const allQueries: ServicesContextProps = {
     ...queriesMock,
+    getAppMetadata,
     ...queries,
   };
 
   return rtlRender(
     <ServicesContextProvider {...allQueries} client={queryClient}>
-      <AccessControlTab {...props} />
+      <AccessControlTab {...defaultProps} {...props} />
     </ServicesContextProvider>,
   );
 };
